requirements.txt
torch>=2.2
torchvision>=0.17
numpy>=1.24
Pillow>=10.0
tqdm>=4.66
scikit-learn>=1.3
PyYAML>=6.0
